requirements/base.txt
Flask
gunicorn
# bravado
# bravado-core
zope.event
connexion[swagger-ui]
python-dotenv
docker-compose